Code cleanup: App Delegate, Login VC
